requirements.txt
streamlit
pandas
numpy
scikit-learn
google-generativeai
PyPDF2
